Move net-worker entrypoint
