+++ src/index.prod.njk
@@ -9,7 +9,7 @@
     <link href="./fonts.css" rel="stylesheet">
     <link href="./base-styles.css" rel="stylesheet">
     <link href="./splash.css" rel="stylesheet">
-    <link rel="prefetch" href="./workers/net-worker/worker.ts" as="worker" />
+    <link rel="prefetch" href="./workers/net-worker.ts" as="worker" />
   </head>
   <body>
     <div id="app"></div>
